update to <1-init>

+++ 0x0E-structures_typedef/1-init_dog.c
@@ -3,7 +3,7 @@
 
 /**
  * init_dog - initializes variable of type struct dog
- * @d: struct
+ * @d: pointer to struct
  * @name: dog name
  * @age: dog age
  * @owner: dog owner
@@ -12,6 +12,6 @@
 void init_dog(struct dog *d, char *name, float age, char *owner)
 {
 	d->name = (char *)name;
-	d->age = age;
+	d->age = (float)age;
 	d->owner = (char *)owner;
 }
